requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.2
pyyaml>=6.0.1
python-dateutil>=2.8.2
rich>=13.7.0
click>=8.1.7
fake-useragent>=1.4.0
lxml>=4.9.3
